requirements.txt
# StockQueen V1 - Python Dependencies
# Generated: 2025-02-25

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
supabase==2.3.0

# HTTP Client
httpx
requests==2.31.0

# RSS Feed Parsing
feedparser==6.0.11

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Scheduling
APScheduler==3.10.4

# SMS Notification
twilio==8.12.0

# Environment Variables
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Validation
pydantic
pydantic-settings

# Date/Time Handling
python-dateutil==2.8.2
pytz==2024.1

# WebSocket Client
websockets==12.0
websocket-client==1.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
